Add files via upload
Committing some changes to files as a backup of the data.  I made changes to all files so that the TileBoard class is now unnecessary and the app will work without it.  I also corrected a few bugs where the main game will not respond if another window within the game is open to prevent weird behavior during gameplay.